+++ src/app/projects/details/page.js
@@ -1,15 +1,11 @@
-"use client";
 import Details from "@/components/details/Details";
-import { useSearchParams } from "next/navigation";
+import { Suspense } from "react";
 
 const Page = () => {
-  const searchParams = useSearchParams();
-  const id = searchParams.get("id");
-
   return (
-    <div>
-      <Details id={id} />
-    </div>
+    <Suspense>
+      <Details />
+    </Suspense>
   );
 };
